Use derive instead of manually implementing serde traits
